pyproject.toml
[tool.poetry]
name = "reflection-agent"
version = "0.1.0"
description = ""
authors = ["nealbffl <neal@bffl.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
black = "^24.8.0"
isort = "^5.13.2"
langchain = "^0.2.15"
langchain-openai = "^0.1.23"
langgraph = "^0.2.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
